+++ server/promptadmin_server/api/service/healthcheck/health_time_service.py
@@ -1,0 +1,9 @@
+import datetime
+
+
+class HealthTimeService:
+    def __init__(self):
+        pass
+
+    def now_index(self):
+        pass