+++ server/promptadmin_server/data/entity/healthcheck/health_target.py
@@ -1,0 +1,8 @@
+from promptadmin_server.commons.entity.base_entity import BaseEntity
+
+
+class HealthTarget(BaseEntity, table=True):
+    __tablename__ = "pa_health_target"
+
+    url: str
+    title: str